통합테스트 진행하며 발생한 버그 수정 (#56)
* fix: 좋아요 취소 예외처리

* fix: querydsl에 is_deleted 조건문 추가

* fix: cursor paging의 jpa sql 수정

* fix: explore 커서페이징 적용

* fix: like exception 수정

* fix: delete like exception

+++ src/main/java/org/ahpuh/surf/like/service/LikeServiceImpl.java
@@ -3,11 +3,14 @@ package org.ahpuh.surf.like.service;
 import lombok.RequiredArgsConstructor;
 import org.ahpuh.surf.common.exception.EntityExceptionHandler;
 import org.ahpuh.surf.like.converter.LikeConverter;
+import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.like.repository.LikeRepository;
 import org.ahpuh.surf.post.entity.Post;
 import org.ahpuh.surf.post.repository.PostRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -28,7 +31,11 @@ public class LikeServiceImpl implements LikeService {
     }
 
     @Override
-    public void unlike(final Long likeId) {
+    public void unlike(final Long postId, final Long likeId) {
+        final Like like = likeRepository.findById(likeId).orElseThrow(() -> new IllegalArgumentException("좋아요한 기록이 없습니다." + likeId));
+        if (!Objects.equals(like.getPost().getPostId(), postId)) {
+            throw new IllegalArgumentException("The post ID does not match. " + postId);
+        }
         likeRepository.deleteById(likeId);
     }
